+++ packages/editor/src/plugins/text-area-exercise/renderer.tsx
@@ -1,9 +1,4 @@
-import { useRef, useState } from 'react'
-
 import { Blocks } from './blocks'
-import { createFeedbackBlock } from './feedback-block'
-import { StateContext } from './state-context'
-import { createTextBlock } from './text-block'
 
 // Schüly Ansicht
 export function TextAreaExerciseRenderer({
@@ -11,16 +6,8 @@ export function TextAreaExerciseRenderer({
 }: {
   scaffoldingEnabled: boolean
 }) {
-  // Global state
-  /*const [state, setState] = useState({
-    blocks: [createTextBlock()],
-    // silentMode: boolean,
-    // ...
-  })*/
-
   return (
     <>
-      {/* Like global state for all child components to use */}
       <Blocks />
       <div>Scaffolding {scaffoldingEnabled ? '✅' : '❌'}</div>
     </>
