--- packages/editor/src/plugins/text-area-exercise/state-context.ts
@@ -1,17 +1,0 @@
-import { createContext } from 'react'
-
-import { Block } from './types'
-
-// Hack: Make context have a type always
-export const StateContext = createContext<State>(null as State)
-
-export interface State {
-  state: {
-    blocks: Block[]
-  }
-  setState: React.Dispatch<
-    React.SetStateAction<{
-      blocks: Block[]
-    }>
-  >
-}
