+++ packages/editor/src/plugins/text-area-exercise/types.ts
@@ -10,3 +10,8 @@ export interface Text {
   type: 'text'
   content: string
 }
+
+export interface AiFeedback {
+  isCorrect: boolean
+  generalFeedback: string
+}
